* io.c (nogvl_io_cntl): rb_cloexec_fcntl_dupfd's 2nd argument is int.
* process.c (move_fds_to_avoid_crash): ditto.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@33602 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ io.c
@@ -7842,7 +7842,7 @@ static VALUE nogvl_io_cntl(void *ptr)
     else
 #if defined(F_DUPFD)
         if (arg->cmd == F_DUPFD)
-            return (VALUE)rb_cloexec_fcntl_dupfd(arg->fd, arg->narg);
+            return (VALUE)rb_cloexec_fcntl_dupfd(arg->fd, (int)arg->narg);
         else
 #endif
             return (VALUE)fcntl(arg->fd, arg->cmd, arg->narg);
